Remove component from package names
