+++ secure-random/src/nativeMain/kotlin/io/matthewnelson/secure/random/internal/-Cinterop.kt
@@ -1,0 +1,22 @@
+/*
+ * Copyright (c) 2023 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ **/
+package io.matthewnelson.secure.random.internal
+
+import kotlinx.cinterop.toKStringFromUtf8
+import platform.posix.strerror
+
+internal fun errnoToString(errno: Int): String = strerror(errno)?.toKStringFromUtf8() ?: "errno: $errno"
